test: Move creds related to tests to creds_test.go
Addresses a TODO in the code and reduces the size of end2end_test.go
ever so slightly.
